Unifico controlador de vistas. No tiene sentido tener OperariosController, además que había conflicto con el nombre de las funciones de controlador. 'home' estaba repetido. https://github.com/Arquisoft/Inci_e3b_modules/issues/29
