+++ src/main/java/es/uniovi/asw/e3b/incidashboard_e3b/controllers/HomeController.java
@@ -1,5 +1,7 @@
 package es.uniovi.asw.e3b.incidashboard_e3b.controllers;
 
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -9,7 +11,21 @@ import org.springframework.web.bind.annotation.RequestMethod;
 public class HomeController {
 	
 	@RequestMapping(value = "/", method = RequestMethod.GET)
-	public String home(Model model) {
+	public String index(Model model) {
 		return "redirect:/login";
-	}	
+	}
+
+	@RequestMapping(value = "/login", method = RequestMethod.GET)
+	public String login(Model model) {
+		return "login";
+	}
+
+
+	@RequestMapping(value = { "/home" }, method = RequestMethod.GET)
+	public String home(Model model) {
+		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+		String email = auth.getName();
+		model.addAttribute("email", email);
+		return "index";
+	}
 }
