--- src/main/java/es/uniovi/asw/e3b/incidashboard_e3b/controllers/OperariosController.java
@@ -1,27 +1,0 @@
-package es.uniovi.asw.e3b.incidashboard_e3b.controllers;
-
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-
-@Controller
-public class OperariosController {
-	
-	@RequestMapping(value = "/login", method = RequestMethod.GET)
-	public String login(Model model) {
-		return "login";
-	}
-	
-	
-	@RequestMapping(value = { "/home" }, method = RequestMethod.GET)
-	public String home(Model model) {
-		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
-		String email = auth.getName();
-		model.addAttribute("email", email);
-	    return "index";
-	}
-
-}
